wip: working for demo - refactored
