+++ src/contexts/ThemeProvider.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useEffect, useState } from 'react'
+import { createContext, useEffect, useState } from 'react'
 
 type Theme = 'dark' | 'light' | 'system'
 
@@ -8,7 +8,7 @@ interface ThemeProviderProps {
   storageKey?: string
 }
 
-interface ThemeProviderState {
+export interface ThemeProviderState {
   theme: Theme
   setTheme: (theme: Theme) => void
 }
@@ -18,7 +18,7 @@ const initialState: ThemeProviderState = {
   setTheme: () => null
 }
 
-const ThemeProviderContext = createContext<ThemeProviderState>(initialState)
+export const ThemeProviderContext = createContext<ThemeProviderState>(initialState)
 
 export const ThemeProvider = ({
   children,
@@ -58,12 +58,4 @@ export const ThemeProvider = ({
       {children}
     </ThemeProviderContext.Provider>
   )
-}
-
-export const useTheme = (): ThemeProviderState => {
-  const context = useContext(ThemeProviderContext)
-
-  if (context === undefined) { throw new Error('useTheme must be used within a ThemeProvider') }
-
-  return context
 }
