+++ src/views/Large.tsx
@@ -1,4 +1,4 @@
-export function Large ({ id }: { id: string }): JSX.Element {
+export function Large({ id }: { id: string }): JSX.Element {
   return (
     <div className="w-[915px] bg-green-200 overflow-hidden">
       <h1 className='font-bold text-lg py-4'>{`Large - ${Date.now()}`}</h1>
