+++ src/views/Small.tsx
